--- src/Components/App.js
@@ -1,10 +1,0 @@
-import React from 'react'
-
-
-export default function App() {
-    return (
-        <div>
-     
-        </div>
-    )
-}
