--- src/Components/Login.js
@@ -1,17 +1,0 @@
-import React from 'react'
-
-
-export default function Login() {
-    return (
-            
-
-            <div className="input">
-                <input type="text" placeholder="Username"/>
-                <input type="text" placeholder="Password"/>
-                <button>Login</button>
-                <p>Please Login with your username and Password</p>
-
-            </div>
-        
-    )
-}
